requirements.txt
django
djangorestframework
dr-buster
drf_yasg
